refactor: Move moltin service outside of utils
