+++ src/components/Cart/CartCounter.js
@@ -4,7 +4,7 @@ import { push } from 'react-router-redux';
 
 import { FETCH_CART_START, FETCH_CART_END } from '../../reducers/cart';
 
-var api = require('../../utils/moltin.js');
+var api = require('../../moltin.js');
 
 function mapStateToProps(state) {
   return state;
