+++ src/components/Cart/CartItems.js
@@ -3,7 +3,7 @@ import { connect } from 'react-redux';
 import ProductImage from '../Products/ProductImage';
 import { FETCH_CART_START, FETCH_CART_END } from '../../reducers/cart';
 
-var api = require('../../utils/moltin.js');
+var api = require('../../moltin.js');
 
 function mapStateToProps(state) {
   return state;
