+++ src/components/Products/ProductsContainer.js
@@ -11,7 +11,7 @@ import {
   FETCH_PRODUCTS_END
 } from '../../reducers/products';
 
-var api = require('../../utils/moltin.js');
+var api = require('../../moltin.js');
 
 const mapStateToProps = state => {
   return {
