+++ src/components/Products/SingleProductContainer.js
@@ -12,7 +12,7 @@ import {
   FETCH_PRODUCTS_END
 } from '../../reducers/products';
 
-var api = require('../../utils/moltin.js');
+var api = require('../../moltin.js');
 
 function mapStateToProps(state) {
   return state;
